feat: add additional record types
